Game Project
article I am showing you how to create a very simple game of Tic-Tac-Toe in a  Python console application.


Introduction

Tic-Tac-Toe is a very simple two player game. So only two players can play at a time. This game is also known as Noughts and Crosses or Xs and Os game. One player plays with X and the other player plays with O. In this game we have a board consisting of a 3X3 grid. The number of grids may be increased.
 
The Tic-Tac-Toe board looks like the following:

  
Game Rules 
1.	Traditionally the first player plays with "X". So you can decide who wants to go with "X" and who wants go with "O".
2.	Only one player can play at a time.
3.	If any of the players have filled a square then the other player and the same player cannot override that square.
4.	There are only two conditions that may match will be draw or may win.
5.	The player that succeeds in placing three respective marks (X or O) in a horizontal, vertical or diagonal row wins the game.
Winning condition

Whoever places three respective marks (X or O) horizontally, vertically or diagonally will be the winner.